+++ src/store/reducer.ts
@@ -2,23 +2,27 @@ import * as actionTypes from "./actionTypes"
 import { v4 as uuidv4 } from 'uuid';
 import { combineReducers } from "redux";
 
-const initialState: TaskState = {
-    tasks: [],
-    editedTask: undefined,
-    searchedTask: '',
+const initialState: State = {
+    taskReducer: {
+        tasks: [],
+        editedTask: undefined,
+    },
+    searchReducer: {
+        searchedTask: '',
+    }
 }
 
 export const taskReducer = (
-    state: TaskState = initialState,
+    state: TaskState = initialState?.taskReducer,
     action: TaskAction
 ): TaskState => {
     switch (action.type) {
         case actionTypes.ADD_TASK:
             const newTask: ITask = { ...action.task, id: uuidv4(), }
 
-            return { ...state, tasks: state.tasks.concat(newTask), }
+            return { ...state, tasks: state?.tasks?.concat(newTask), }
         case actionTypes.UPDATE_TASK:
-            const editedTasks: ITask[] = state.tasks.map(
+            const editedTasks: ITask[] = state?.tasks?.map(
                 task => {
                     if (task.id !== action.task.id) return task;
 
@@ -28,14 +32,14 @@ export const taskReducer = (
 
             return { ...state, tasks: editedTasks, }
         case actionTypes.REMOVE_TASK:
-            const updatedTasks: ITask[] = state.tasks.filter(
+            const updatedTasks: ITask[] = state?.tasks.filter(
                 task => task.id !== action.task.id
             )
 
             return { ...state, tasks: updatedTasks, }
         case actionTypes.SET_EDITED_TASK:
             const targetId = action.task.id;
-            const taskToEdit: ITask|undefined = state.tasks.find(
+            const taskToEdit: ITask|undefined = state?.tasks?.find(
                 ({ id }) => id === targetId
             );
 
@@ -47,10 +51,12 @@ export const taskReducer = (
     return state
 }
 
+// TODO: create edit reducer
+
 export const searchReducer = (
-    state: TaskState = initialState,
+    state: SearchState = initialState?.searchReducer,
     action: SearchAction
-): TaskState => {
+): SearchState => {
     switch (action.type) {
         case actionTypes.SET_SEARCHED_TASK:
             return { ...state, searchedTask: action.value };